💥 Remove Amplify
