+++ package.json
@@ -9,8 +9,6 @@
     "snyk-protect": "snyk protect"
   },
   "dependencies": {
-    "@aws-amplify/auth": "^3.2.6",
-    "@aws-amplify/core": "^3.2.6",
     "@hapi/joi": "^17.1.1",
     "autoprefixer": "^9.8.0",
     "aws-sdk": "^2.675.0",
