+++ pages/api/user/[userID]/index.ts
@@ -6,7 +6,7 @@ import {
   getUserByUsername,
   updateUser,
 } from "util/db/user";
-import { checkAuth } from "util/auth";
+import { checkAuth } from "util/authServer";
 
 const index = async (req: NextApiRequest, res: NextApiResponse) => {
   try {
